Created metrics.py, moved all metrics from suitability.py to metrics.py, created compare_matrices() function to run a comparison of the LLM based suitability matrix and the rule based suitability matrix and determine how well the LLM performs.
